requirements.txt
tensorflow==2.10.0
numpy==1.23.5
pandas==1.3.5
matplotlib==3.5.3
opencv-python==4.6.0.66
Pillow==9.2.0
scikit-learn==1.0.2